INT-3384:change code for Artem's comments
