+++ spring-integration-jms/src/main/java/org/springframework/integration/jms/JmsOutboundGateway.java
@@ -48,7 +48,7 @@ import org.springframework.integration.IntegrationMessageHeaderAccessor;
 import org.springframework.integration.MessageTimeoutException;
 import org.springframework.integration.handler.AbstractReplyProducingMessageHandler;
 import org.springframework.integration.handler.ExpressionEvaluatingMessageProcessor;
-import org.springframework.integration.jms.config.JmsAdapterParserUtils;
+import org.springframework.integration.jms.util.JmsAdapterUtils;
 import org.springframework.jms.connection.ConnectionFactoryUtils;
 import org.springframework.jms.listener.DefaultMessageListenerContainer;
 import org.springframework.jms.support.JmsUtils;
@@ -595,9 +595,9 @@ public class JmsOutboundGateway extends AbstractReplyProducingMessageHandler imp
 				container.setRecoveryInterval(this.replyContainerProperties.getRecoveryInterval());
 			}
 			if (this.replyContainerProperties.getSessionAcknowledgeMode() != null) {
-				Integer acknowledgeMode = JmsAdapterParserUtils.parseAcknowledgeMode(this.replyContainerProperties.getSessionAcknowledgeMode());
+				Integer acknowledgeMode = JmsAdapterUtils.parseAcknowledgeMode(this.replyContainerProperties.getSessionAcknowledgeMode());
 				if (acknowledgeMode != null) {
-					if (JmsAdapterParserUtils.SESSION_TRANSACTED == acknowledgeMode) {
+					if (JmsAdapterUtils.SESSION_TRANSACTED == acknowledgeMode) {
 						container.setSessionTransacted(true);
 					}
 				}
