+++ spring-integration-jms/src/main/java/org/springframework/integration/jms/config/JmsInboundChannelAdapterParser.java
@@ -63,7 +63,7 @@ public class JmsInboundChannelAdapterParser extends AbstractPollingInboundChanne
 		}
 		else {
 			builder.addConstructorArgValue(JmsAdapterParserUtils.parseJmsTemplateBeanDefinition(element, parserContext));
-			IntegrationNamespaceUtils.setValueIfAttributeDefined(builder, element, "acknowledge","sessionAcknowledgeMode");
+			IntegrationNamespaceUtils.setValueIfAttributeDefined(builder, element, "acknowledge", "sessionAcknowledgeMode");
 		}
 		if (hasDestinationRef || hasDestinationName) {
 			if (hasDestinationRef) {
