+++ spring-integration-xml/src/main/resources/org/springframework/integration/xml/config/spring-integration-xml-4.1.xsd
@@ -838,12 +838,14 @@
 			<xsd:attribute name="expression" type="xsd:string" use="optional"/>
 			<xsd:attribute name="value" type="xsd:string" use="optional"/>
 	</xsd:complexType>
+	
 	<xsd:simpleType name="resultType">
 		<xsd:restriction base="xsd:string">
 			<xsd:enumeration value="DOMResult"/>
 			<xsd:enumeration value="StringResult"/>
 		</xsd:restriction>
 	</xsd:simpleType>
+	
 	<xsd:simpleType name="evaluationType">
 		<xsd:restriction base="xsd:string">
 		<xsd:enumeration value="BOOLEAN_RESULT"/>
@@ -853,10 +855,12 @@
 		<xsd:enumeration value="NODE_LIST_RESULT"/>
 		</xsd:restriction>
 	</xsd:simpleType>
+	
 	<xsd:simpleType name="schemaType">
 		<xsd:restriction base="xsd:string">
 		<xsd:enumeration value="xml-schema"/>
 		<xsd:enumeration value="relax-ng"/>
 		</xsd:restriction>
 	</xsd:simpleType>
+	
 </xsd:schema>
