+++ spring-integration-xml/src/test/java/org/springframework/integration/xml/config/XPathTransformerParserTests-context.xml
@@ -13,7 +13,9 @@
 			http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd
 			http://www.springframework.org/schema/integration/xml
 			http://www.springframework.org/schema/integration/xml/spring-integration-xml.xsd">
+			
 	<context:property-placeholder properties-ref="props"/>
+	
 	<util:properties id="props">
 		<beans:prop key="booleanResult">BOOLEAN_RESULT</beans:prop>
 		<beans:prop key="stringResult">STRING_RESULT</beans:prop>
@@ -21,6 +23,7 @@
 		<beans:prop key="nodeResult">NODE_RESULT</beans:prop>
 		<beans:prop key="nodeListResult">NODE_LIST_RESULT</beans:prop>
  	</util:properties>
+ 	
 	<si:channel id="output">
 		<si:queue/>
 	</si:channel>
