+++ spring-integration-xml/src/test/java/org/springframework/integration/xml/transformer/jaxbmarshaling/JaxbMarshallingIntegrationTests-context.xml
@@ -14,6 +14,7 @@
 			http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd
 			http://www.springframework.org/schema/context 
 			http://www.springframework.org/schema/context/spring-context.xsd">
+			
 	<context:property-placeholder properties-ref="props"/>
 
 	<util:properties id="props">
@@ -26,7 +27,6 @@
 	
 	<si:channel id="unmarshallIn"/>
 		
-		
 	<si:channel id="unmarshallOut">
 		<si:queue capacity="10" />
 	</si:channel>
@@ -37,9 +37,7 @@
 	
 	<si-xml:marshalling-transformer id="marshaller" input-channel="marshallIn" output-channel="marshallOut" marshaller="marshallerUnmarshaller" result-type="${result}" />
 	
-	
 	<si-xml:unmarshalling-transformer id="unmarshaller" input-channel="unmarshallIn" output-channel="unmarshallOut" unmarshaller="marshallerUnmarshaller" />
-	
 	
 	<bean id="marshallerUnmarshaller" class="org.springframework.oxm.jaxb.Jaxb2Marshaller" >
 		<property name="classesToBeBound">
